[#16] refactor: DTO 패키지 변경
