+++ src/main/java/dabang/star/cafe/api/exception/CustomExceptionHandler.java
@@ -1,6 +1,5 @@
 package dabang.star.cafe.api.exception;
 
-import dabang.star.cafe.api.response.ErrorResponse;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.MethodArgumentNotValidException;
