+++ src/main/java/dabang/star/cafe/domain/login/LoginService.java
@@ -1,8 +1,8 @@
 package dabang.star.cafe.domain.login;
 
 
-import dabang.star.cafe.api.response.manager.ManagerData;
-import dabang.star.cafe.api.response.member.MemberData;
+import dabang.star.cafe.domain.manager.ManagerData;
+import dabang.star.cafe.domain.member.MemberData;
 
 public interface LoginService {
