+++ src/main/java/dabang/star/cafe/domain/manager/ManagerRepository.java
@@ -1,7 +1,5 @@
 package dabang.star.cafe.domain.manager;
 
-import dabang.star.cafe.api.response.manager.ManagerData;
-
 import java.util.Optional;
 
 public interface ManagerRepository {
