+++ src/main/java/dabang/star/cafe/domain/member/MemberRepository.java
@@ -1,7 +1,5 @@
 package dabang.star.cafe.domain.member;
 
-import dabang.star.cafe.api.response.member.MemberData;
-
 import java.util.Optional;
 
 public interface MemberRepository {
