+++ src/main/java/dabang/star/cafe/domain/member/MemberService.java
@@ -1,7 +1,5 @@
 package dabang.star.cafe.domain.member;
 
-import dabang.star.cafe.api.response.member.MemberData;
-
 public interface MemberService {
 
     MemberData join(Member member);
