+++ src/main/java/dabang/star/cafe/infrastructure/mapper/ManagerMapper.java
@@ -1,6 +1,6 @@
 package dabang.star.cafe.infrastructure.mapper;
 
-import dabang.star.cafe.api.response.manager.ManagerData;
+import dabang.star.cafe.domain.manager.ManagerData;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Param;
