+++ src/main/java/dabang/star/cafe/infrastructure/mapper/MemberMapper.java
@@ -1,6 +1,6 @@
 package dabang.star.cafe.infrastructure.mapper;
 
-import dabang.star.cafe.api.response.member.MemberData;
+import dabang.star.cafe.domain.member.MemberData;
 import dabang.star.cafe.domain.member.Member;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Param;
